+++ src/main/kotlin/waig548/XNN/internal/network/NetworkBase.kt
@@ -1,13 +1,15 @@
 package waig548.XNN.internal.network
 
+import java.math.BigDecimal
+
 sealed class NetworkBase
 {
     abstract val size: Int
 
-    var x: List<Double> = listOf()
-    var y: List<Double> = listOf()
-    var output: List<Double> = listOf()
+    var x: List<BigDecimal> = listOf()
+    var y: List<BigDecimal> = listOf()
+    var output: List<BigDecimal> = listOf()
 
-    //abstract fun forward(input: List<Double>): List<Double>
-    //abstract fun backprop(desired: List<Double>): List<Double>
+    //abstract fun forward(input: List<BigDecimal>): List<BigDecimal>
+    //abstract fun backprop(desired: List<BigDecimal>): List<BigDecimal>
 }
